Module-out `getInputs`
